# smartlock-docs/enterprise/16_API_Spec.yaml
...
openapi: 3.1.0
info:
  title: Smart Lock AI Support & Service Dispatch SaaS — 同步介面契約
  version: "1.1"
  summary: 平台所有同步 HTTP / WebSocket 介面的單一契約入口
  description: |
    本檔涵蓋平台 6 子系統的同步介面：

    | 介面 | 提供者 → 消費者 |
    |---|---|
    | 品牌營運控制平面（工單/派工/結算/帳務/治理；runtime export 517 端點）| api → web(dispatch) / 平台 console |
    | agent 服務間旁路 `/internal/*`（4 端點）| agent → api |
    | LINE 入站 webhook `POST /callback` | LINE 平台 → agent（見 webhooks 區）|
    | OHS API（技師查詢/媒合/排班/認證）| technician-platform → 各品牌 api 消費 |
    | knowledge-refinery 審核佇列 API | refinery web → refinery svc（摘要，見 refinery tag）|

    **組織原則**：runtime 517 端點不逐一展開；以 tags + 各領域代表性資源組織，其餘端點見各
    tag `externalDocs` 指向的系統契約文件（完整端點總表 `../api/P2/06` §3）。

    **通用約定**（`../api/P2/06` §1）：
    - 版本切分：v1（legacy）掛 `/api/v1` 前綴且所有回應帶 `Deprecation: true` header；
      v2（tenant-scoped）走 `/tenants/{tenantId}/...` flat path。
    - 認證：browser HTTP 走本站 proxy + HttpOnly access/refresh cookie；CLI/mobile 可用
      Casdoor OIDC bearer。服務間優先 `X-Service-Credential`，`X-Internal-Token` 只作
      CR-0190 migration fallback。
    - 回應信封：成功 `{data, error: null}`（ApiResponse）/ 游標分頁 `CursorPage`；
      錯誤統一 RFC7807 superset（`application/problem+json`，見 Problem schema）。
    - 冪等：寫入類支援 `Idempotency-Key` header，TTL 24h。
    - SoD：雙簽/三簽端點需 `X-Initiator` / `X-Approver`（+ 選填 `X-Executor`），
      任二相同 → 403 SOD_VIOLATION。
    - 即時推播：WS 頻道契約見 `x-websocket-channels`；fanout 傳輸設計見 ./17_AsyncAPI.yaml。
...
      內部 MCP 協議，非 HTTP，不進本契約）。
  contact: { name: 平台架構組 }
...
servers:
...
security:
  - casdoorOIDC: []
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: 品牌管理員登入（另有 login-vendor / technician / platform 登入、refresh、logout）
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/LoginRequest" } } }
      responses:
        "200": { description: 登入成功回 token 組, content: { application/json: { schema: { $ref: "#/components/schemas/TokenPair" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ── B. 工單（完整代表性資源）──────────────────────────────────────────
...
  /internal/conversations/ingest:
    post:
      tags: [internal]
      summary: 旁路持久化一輪 LINE 對話（回覆後 fire-and-forget；agent 側 timeout 20s）
      security: [{ internalToken: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/IngestTurnRequest" } } }
      responses:
        "200": { description: 已寫入 conversations/messages, content: { application/json: { schema: { $ref: "#/components/schemas/ApiResponse" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /internal/conversations/handover-state:
    get:
      tags: [internal]
      summary: 查對話是否人工接管中（agent 回覆前判斷；timeout 5s，agent 側 fail-soft 回 false）
      security: [{ internalToken: [] }]
      parameters:
        - { name: tenant_id, in: query, required: true, schema: { type: string } }
        - { name: session_id, in: query, required: true, schema: { type: string }, description: "`{tenant}:{line_user_id}`" }
      responses:
        "200": { description: 接管狀態（查無 session → escalated=false）, content: { application/json: { schema: { $ref: "#/components/schemas/HandoverState" } } } }
  /internal/escalations/ingest:
    post:
      tags: [internal]
      summary: agent transfer_to_human → 建 AI 草擬問題卡（source='ai_line'）
      security: [{ internalToken: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/EscalationIngestRequest" } } }
      responses:
        "201": { description: 草擬卡已建立（confirm/convert 走客服認證端點）, content: { application/json: { schema: { $ref: "#/components/schemas/ApiResponse" } } } }
  /internal/quotes/{quoteId}:customer-respond:
    post:
      tags: [internal]
      summary: 客戶經 LINE postback 同意 / 拒絕報價（先驗客戶身分，防跨客戶誤同意）
      security: [{ internalToken: [] }]
      parameters:
        - { name: quoteId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/QuoteCustomerRespond" } } }
      responses:
        "200": { description: 已記錄客戶決定, content: { application/json: { schema: { $ref: "#/components/schemas/ApiResponse" } } } }
...
components:
  securitySchemes:
    casdoorOIDC:
      type: openIdConnect
      openIdConnectUrl: https://casdoor.platform.internal/.well-known/openid-configuration  # [待確認] 正式 Casdoor URL
      description: |
        Casdoor 統一 IdP（ADR-P003）。各服務驗 OIDC bearer token；角色由 role claim
        提供，四方 RBAC（Super Admin / 租戶 Admin / 派工小編 / 技師跨租戶身分）
        deny-by-default enforce（ADR-P006）。守衛鏈：require_tenant（X-Tenant-ID 與
        claim 一致）→ role_required(*roles) → require_platform_admin（跨品牌視角）
        → require_sod_actors / require_keeper_role（特權操作）。
        系統角色 12 個：admin / reviewer / technician / brand_oem / line_user /
        accounting / supervisor / dispatcher / customer_service / auditor /
        family_reviewer / distributor。
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: CR-0190 migration fallback only；production usage 須連續一個 release window 為零後移除。
    serviceCredential:
      type: apiKey
      in: header
      name: X-Service-Credential
      description: Per-workload opaque credential；hash-only、audience/scope/tenant/expiry/revoke/audit。
    ohsServiceCredential:
      type: apiKey
      in: header
      name: X-Service-Credential
      description: "品牌 api → technician-platform OHS 現行可用的受控 opaque credential；OD-001 仍決定最終是否改 OIDC client-credentials。"
    wsCookieAuth:
      type: apiKey
      in: cookie
      name: smartlock_access_token
      description: WS/SSE 認證只走 HttpOnly cookie；URL 不得含 access/refresh token。跨 host 無共享 cookie 時 realtime disabled。
...
  responses:
...
    Unauthorized:
      description: 缺/無效/過期 token（UNAUTHENTICATED、TOKEN_REVOKED、TOKEN_STALE、INTERNAL_AUTH_FAILED）
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
    ServiceUnavailable:
      description: 內部憑證未配置（INTERNAL_AUTH_NOT_CONFIGURED，fail-closed）
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
  schemas:
...
    Problem:
      type: object
      description: "RFC7807 superset 錯誤信封；type 為 urn:smartlock:error:{error_code_lowercase}（識別字串非可抓取 URL）；error_code/message/request_id 為 extension members"
      required: [type, title, status, error_code, request_id]
      properties:
        type: { type: string, examples: ["urn:smartlock:error:validation_error"] }
        title: { type: string, examples: ["Validation Error"] }
        status: { type: integer, examples: [422] }
        detail: { type: string }
        instance: { type: string }
        error_code: { type: string, examples: [VALIDATION_ERROR] }
        message: { type: string }
        request_id: { type: string }
        timestamp: { type: string, format: date-time }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string, examples: [body.amount] }
              issue: { type: string }
              type: { type: string }
    ApiResponse:
      type: object
      description: 成功回應信封
      properties:
        data: {}
        error: { type: "null" }
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    TokenPair:
      type: object
      properties:
        access_token: { type: string, description: Casdoor OIDC access token }
        refresh_token: { type: string }
        token_type: { type: string, const: Bearer }
        expires_in: { type: integer }
...
    IngestTurnRequest:
      type: object
      required: [tenant_id, line_user_id, session_id, user_text, assistant_text]
      properties:
        tenant_id: { type: string, examples: [locksmart] }
        line_user_id: { type: string }
        session_id: { type: string, description: "`{tenant}:{line_user_id}`" }
        user_text: { type: string }
        assistant_text: { type: string }
        display_name: { type: string }
    HandoverState:
      type: object
      properties:
        data:
          type: object
          properties:
            escalated: { type: boolean }
        error: { type: "null" }
    EscalationIngestRequest:
      type: object
      required: [tenant_id, line_user_id, session_id, reason]
      properties:
        tenant_id: { type: string }
        line_user_id: { type: string }
        session_id: { type: string }
        reason: { type: string }
        is_explicit: { type: boolean, description: 金錢 / 明確要真人關鍵字偵測 }
        facts_snapshot:
          type: object
          properties:
            facts_block: { type: string }
            user_input_excerpt: { type: string }
            brand: { type: string }
            model: { type: string }
            symptom: { type: string }
            phone: { type: string, description: 空白時由本輪原話補抽台灣手機 }
    QuoteCustomerRespond:
      type: object
      required: [tenant_id, line_user_id, decision]
      properties:
        tenant_id: { type: string }
        line_user_id: { type: string }
        decision: { type: string, enum: [accept, reject] }
...